LPS-62297 moved to portal-settings
